# backend-1/src/open-api/swagger.yaml
openapi: 3.0.0
info:
  title: IoT API
  version: 1.0.0
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /api/maintenance:
    get:
      summary: Get maintenance information
      description: >-
        Returns maintenance links, digital twin info, and IODD catalog. Data is
        cached in Redis.
      tags:
        - Maintenance
      responses:
        '200':
          description: Maintenance data including links and IODD catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  digitaltwin:
                    type: object
                    properties:
                      forms:
                        type: array
                        items:
                          type: object
                  backup_restore:
                    type: object
                  factorysettings:
                    type: object
                  firmware_update:
                    type: object
                  iodd_catalog:
                    type: object
                    properties:
                      ethIpv4:
                        type: array
                        items:
                          type: object
        '500':
          description: Internal server error
    put:
      summary: Update maintenance data
      description: Updates maintenance configuration. Requires authentication.
      tags:
        - Maintenance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated maintenance data
        '404':
          description: Maintenance not found
        '500':
          description: Failed to update maintenance
  /api/maintenance/reboot:
    post:
      summary: Reboot device
      description: Triggers device reboot. Requires authentication.
      tags:
        - Maintenance
      responses:
        '200':
          description: Reboot initiated successfully
        '500':
          description: Failed to initiate reboot
  /api/maintenance/factorysettings:
    post:
      summary: Reset to factory settings
      description: Resets device to factory settings. Requires authentication.
      tags:
        - Maintenance
      responses:
        '200':
          description: Factory reset initiated successfully
        '500':
          description: Failed to initiate factory reset
  /api/maintenance/fwupdate:
    post:
      summary: Update firmware
      description: Initiates firmware update process. Requires authentication.
      tags:
        - Maintenance
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                firmware:
                  type: string
                  format: binary
      responses:
        '200':
          description: Firmware update initiated successfully
        '400':
          description: Invalid firmware file
        '500':
          description: Failed to initiate firmware update
...